Added for Bid post and get latest bid

+++ src/main/java/com/sapient/auction/sale/repository/SaleRepository.java
@@ -19,5 +19,5 @@ public interface SaleRepository {
 
 	boolean bid(Bid bid);
 
-	Bid getLatestBid(int saleId);
+	Bid getLatestBid(Long saleId);
 }
